xbps: bump to depend on glibc>=2.25_1
We have versioned symbols for glibc, if you update an old system with
the previous glibc release installed, only xbps is updated and breaks
because of those versioned symbols.
Increasing the required glibc version in common/shlibs and bumping xbps
fixes this and would update glibc too if it updates itself.

https://forum.voidlinux.eu/t/update-of-fresh-install-breaking-xbps/2143