gajim: set python interpreter to python2 in configure_args (#6143)
Gajim uses shell scripts as wrappers to start the actual application. It detects the python interpreter during the configure step and writes the absolute path to the interpreter in the scripts. Since these scripts are shell scripts and the python interpreter is not in the shebang line, it is not caught by the shebang rewrite of xbps-src. This can lead to situations where /usr/bin/python is in the wrapper scripts and is linked to python 3 so the start of gajim fails.

Closes #6099.