hostapd: Add support for logging to syslog (#6174)
Even if enabled in hostapd.conf the 2.6 release only logs some of its
messages to syslog. All parts of the code using the wpa_printf function
send output to stdout.

This commit adds a patch (taken from hostap.git) to the package to fully
support logging to syslog using the -s commandline argument. The runit
run file is modified accordingly.