findutils: remove locate.
Findutils locate/updatedb is very inefficient, and either an information
leak or not useful, since the database is public readable.  It's slocate
implementation warns it has bugs.

Providing this locate implementation was deemed impractical, especially
in a core package like findutils.

Users interested in having a proper locate service are highly
recommended to use mlocate instead, which has none of these
shortcommings.

Closes #7032.