gegl: revert to 0.2.0
Instead of rebuilding gimp im going to revert this for the following
reasons:

Because the shared library provided by gegl is not versioned this
package has to stay with 0.2 as long as there is another package
depending on it otherwise we end up with broken dependencies,
which in this case blocked a bunch of packages from entering the
repository.

There is already another gegl3 package, adding a new gegl4 package
is probably the best way for now.

cc #14063