New package: strawberry-0.4.2
This is a fork of Clementine based on Qt5. It supports roughly the
same featureset and has the same UI, except it uses Qt5 instead of
Qt4 and certain features are different (e.g. no Spotify support,
but there is Tidal streaming support).

For now, we are building without Deezer support (missing support
packages in repo).