cross-aarch64-linux-gnu: fix incorrect substitutions
The lib64 subst was copy pasted from the first one, but only
partially modified. And libpthread.so is a real shared library,
or a symlink to one, so it should not be modified.

Fortunately, according to my builds, nothing in libpthread.so was
ever matched, so there was no harm done. And the existing lib64
match already did the trick for that particular case, but fix
it anyway, for consistency.

This does not result in any difference in the binary build, but
it does fix the correctness.