mirrorbits: update to 0.5.1.
This requires some trickery due to mirrobits depending on an older
version of the protoc-gen-go package. Because the Protobuf originally
generated for the project isn't included, pre_build needs to fetch and
install the correct plugin version, since newer releases are not
compatible with the Protobuf library mirrorbits depends on.

As such, this fetches the github.com/golang/protobuf library at the
commit used by mirrorbits and installs the Protobuf plugin from that.
Once installed, it can generate its the RPC Protobuf code with protoc.

This skips using make in the interest of not patching a makefile.

Closes: #7455 [via git-merge-pr]