New package: libvpx5-1.7.0
This is necessary in order to transition to newer libvpx. Web
browsers, particularly their vendored webrtc component, depend
on libvpx with older API, while new versions in the future will
depend on the version with the newer API. In order to break the
cycle, we need to temporarily provide both versions, and remove
the older version once everything is updated.